added Differential Evolution
very good results on BC and Shaffers, doesn't seem to work at all on
katsuura, maybe with tuning, but unlikely.
Parameters are crossover probability (cr) and intensity of the mutation
(f)
It's possible to decide which base to use for the mutant vectors( 'b'
for best, 'r' for random) and how many couples of parents (diffN).